research-bot-v6/pyproject.toml
[project]
name = "research-bot"
version = "0.1.0"
description = "AI-powered research assistant using LangChain 1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "uvicorn>=0.38.0",
    "pydantic>=2.0.0",
    "python-multipart",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-supervisor>=0.0.30",
    "langmem>=0.0.30",
    "langsmith>=0.1.0",
    "qdrant-client>=1.11.0",
    "openai>=1.0.0",
    "jupyter",
]
